requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-qdrant
qdrant-client
openai
python-dotenv